+++ debug/example/activity/activity.js
@@ -8,12 +8,12 @@ const onReady = () => {
   // #docs/start : hash 경로가 바뀔때 마다 호출 됩니다
   const urlManager = new UrlManager(document.location.toString());
   new ElementButton('#go_right').addEventListener('click', function () {
+    history.pushState('#slideR2L', 'right', `?a=b`);
     const activityId = Activity.onStart('fvueSlideFromRight top-[200px]', 'fvueSlideToLeft');
-    history.pushState('#slideR2L', 'right', `?a=b#/${activityId}`);
   });
   new ElementButton('#go_bottom').addEventListener('click', function () {
+    history.pushState('#slideB2T', 'bottom', `?b=c`);
     const activityId = Activity.onStart('fvueSlideFromBottom top-[140%]', 'fvueSlideToTop');
-    history.pushState('#slideB2T', 'bottom', `#/${activityId}`);
   });
 };
 // document ready