+++ debug/plugins/prism-es6/components/prism-elixir.js
@@ -873,6 +873,10 @@ Prism.languages.elixir = {
 
 
 
+
+
+
+
       // See interpolation below
     } }, { pattern: /("""|''')[\s\S]*?\1/, greedy: true, inside: {