+++ src/example/activity/activity.ts
@@ -13,13 +13,13 @@ const onReady = () : void =>
     const urlManager = new UrlManager(document.location.toString());
 
     new ElementButton('#go_right')!.addEventListener('click', function(){
+        history.pushState('#slideR2L', 'right', `?a=b`);
         const activityId = Activity.onStart('fvueSlideFromRight top-[200px]', 'fvueSlideToLeft');
-        history.pushState('#slideR2L', 'right', `?a=b#/${activityId}`);
     });
 
     new ElementButton('#go_bottom')!.addEventListener('click', function(){
+        history.pushState('#slideB2T', 'bottom', `?b=c`);
         const activityId = Activity.onStart('fvueSlideFromBottom top-[140%]', 'fvueSlideToTop');
-        history.pushState('#slideB2T', 'bottom', `#/${activityId}`);
     });
 };